Removed warnings in windows.

+++ include/microcdr/microcdr.h
@@ -76,9 +76,9 @@ microcdr_DllAPI void restore_micro_state  (MicroBuffer* buffer, const MicroState
 microcdr_DllAPI void align_to(MicroBuffer* buffer, const uint32_t size);
 microcdr_DllAPI uint32_t get_alignment(uint32_t current_alignment, const uint32_t data_size);
 
-microcdr_DllAPI uint32_t micro_buffer_size(const MicroBuffer* buffer);
-microcdr_DllAPI uint32_t micro_buffer_length(const MicroBuffer* buffer);
-microcdr_DllAPI uint32_t micro_buffer_remaining(const MicroBuffer* buffer);
+microcdr_DllAPI size_t micro_buffer_size(const MicroBuffer* buffer);
+microcdr_DllAPI size_t micro_buffer_length(const MicroBuffer* buffer);
+microcdr_DllAPI size_t micro_buffer_remaining(const MicroBuffer* buffer);
 
 // -------------------------------------------------------------------
 //                        SERIALIZATION FUNCTIONS
